+++ test/test_selectors.py
@@ -2,6 +2,7 @@ from __future__ import with_statement
 import errno
 import os
 import psutil
+import select
 import signal
 import socket
 import sys
@@ -49,6 +50,11 @@ TOLERANCE = 0.75
 TRAVIS_CI = "TRAVIS" in os.environ
 
 
+skipUnlessHasSelector = skipUnless(selectors.HAS_SELECT, "Platform doesn't have a selector")
+skipUnlessHasENOSYS = skipUnless(hasattr(errno, 'ENOSYS'), "Platform doesn't have errno.ENOSYS")
+skipUnlessHasAlarm = skipUnless(hasattr(signal, 'alarm'), "Platform doesn't have signal.alarm()")
+
+
 class AlarmThread(threading.Thread):
     def __init__(self, timeout):
         super(AlarmThread, self).__init__(group=None)
@@ -69,8 +75,6 @@ class AlarmMixin(object):
     alarm_thread = None
 
     def _begin_alarm_thread(self, timeout):
-        if not HAS_ALARM:
-            self.skipTest("Platform doesn't have signal.SIGALRM")
         self.addCleanup(self._cancel_alarm_thread)
         self.alarm_thread = AlarmThread(timeout)
         self.alarm_thread.start()
@@ -117,10 +121,9 @@ class TimerMixin(object):
         return TimerContext(self, lower=lower, upper=upper)
 
 
-@skipUnless(selectors.HAS_SELECT, "Platform doesn't have a selector")
+@skipUnlessHasSelector
 class BaseSelectorTestCase(unittest.TestCase, AlarmMixin, TimerMixin):
     """ Implements the tests that each type of selector must pass. """
-    SELECTOR = selectors.DefaultSelector
 
     def make_socketpair(self):
         rd, wr = socketpair()
@@ -135,9 +138,30 @@ class BaseSelectorTestCase(unittest.TestCase, AlarmMixin, TimerMixin):
         return rd, wr
 
     def make_selector(self):
-        s = self.SELECTOR()
+        s = selectors.DefaultSelector()
         self.addCleanup(s.close)
         return s
+
+    def patch_select_module(self, *keep, **replace):
+        """ Helper function that removes all selectors from the select module
+        except those listed in *keep and **replace. Those in keep will be kept
+        if they exist in the select module and those in replace will be patched
+        with the value that is given regardless if they exist or not. Cleanup
+        will restore previous state. This helper also resets the selectors module
+        so that a call to DefaultSelector() will do feature detection again. """
+        selectors._DEFAULT_SELECTOR = None
+        for s in ['select', 'poll', 'epoll', 'kqueue']:
+            if s in replace:
+                if hasattr(select, s):
+                    old_selector = getattr(select, s)
+                    self.addCleanup(setattr, select, s, old_selector)
+                else:
+                    self.addCleanup(delattr, select, s)
+                setattr(select, s, replace[s])
+            elif s not in keep and hasattr(select, s):
+                old_selector = getattr(select, s)
+                self.addCleanup(setattr, select, s, old_selector)
+                delattr(select, s)
 
     def standard_setup(self):
         s = self.make_selector()
@@ -418,7 +442,7 @@ class BaseSelectorTestCase(unittest.TestCase, AlarmMixin, TimerMixin):
         with self.assertTakesTime(lower=SHORT_SELECT, upper=SHORT_SELECT):
             self.assertEqual(0, len(s.select(timeout=SHORT_SELECT)))
 
-    @skipUnless(HAS_ALARM, "Platform doesn't have signal.alarm()")
+    @skipUnlessHasAlarm
     def test_select_timing(self):
         s = self.make_selector()
         rd, wr = self.make_socketpair()
@@ -430,7 +454,7 @@ class BaseSelectorTestCase(unittest.TestCase, AlarmMixin, TimerMixin):
             ready = s.select(LONG_SELECT)
         self.assertEqual([(key, selectors.EVENT_READ)], ready)
 
-    @skipUnless(HAS_ALARM, "Platform doesn't have signal.alarm()")
+    @skipUnlessHasAlarm
     def test_select_interrupt_no_event(self):
         s = self.make_selector()
         rd, wr = self.make_socketpair()
@@ -441,7 +465,7 @@ class BaseSelectorTestCase(unittest.TestCase, AlarmMixin, TimerMixin):
         with self.assertTakesTime(lower=LONG_SELECT, upper=LONG_SELECT):
             self.assertEqual([], s.select(LONG_SELECT))
 
-    @skipUnless(HAS_ALARM, "Platform doesn't have signal.alarm()")
+    @skipUnlessHasAlarm
     def test_select_interrupt_with_event(self):
         s = self.make_selector()
         rd, wr = self.make_socketpair()
@@ -454,7 +478,7 @@ class BaseSelectorTestCase(unittest.TestCase, AlarmMixin, TimerMixin):
             self.assertEqual([(key, selectors.EVENT_READ)], s.select(LONG_SELECT))
         self.assertEqual(rd.recv(1), b'x')
 
-    @skipUnless(HAS_ALARM, "Platform doesn't have signal.alarm()")
+    @skipUnlessHasAlarm
     def test_select_multiple_interrupts_with_event(self):
         s = self.make_selector()
         rd, wr = self.make_socketpair()
@@ -474,7 +498,7 @@ class BaseSelectorTestCase(unittest.TestCase, AlarmMixin, TimerMixin):
             self.assertEqual([(key, selectors.EVENT_READ)], s.select(LONG_SELECT))
         self.assertEqual(rd.recv(1), b'x')
 
-    @skipUnless(HAS_ALARM, "Platform doesn't have signal.alarm()")
+    @skipUnlessHasAlarm
     def test_selector_error(self):
         s = self.make_selector()
         rd, wr = self.make_socketpair()
@@ -498,7 +522,7 @@ class BaseSelectorTestCase(unittest.TestCase, AlarmMixin, TimerMixin):
 
     # Test ensures that _syscall_wrapper properly raises the
     # exception that is raised from an interrupt handler.
-    @skipUnless(HAS_ALARM, "Platform doesn't have signal.alarm()")
+    @skipUnlessHasAlarm
     def test_select_interrupt_exception(self):
         s = self.make_selector()
         rd, wr = self.make_socketpair()
@@ -619,7 +643,7 @@ class BaseWaitForTestCase(unittest.TestCase, TimerMixin, AlarmMixin):
         wait.wait_for_write([rd, wr], 0.001)
         self.assertIs(selector._map, None)
 
-    @skipUnless(HAS_ALARM, "Platform doesn't have signal.alarm()")
+    @skipUnlessHasAlarm
     def test_interrupt_wait_for_read_no_event(self):
         rd, wr = self.make_socketpair()
 
@@ -627,7 +651,7 @@ class BaseWaitForTestCase(unittest.TestCase, TimerMixin, AlarmMixin):
         with self.assertTakesTime(lower=LONG_SELECT, upper=LONG_SELECT):
             self.assertEqual([], wait.wait_for_read(rd, timeout=LONG_SELECT))
 
-    @skipUnless(HAS_ALARM, "Platform doesn't have signal.alarm()")
+    @skipUnlessHasAlarm
     def test_interrupt_wait_for_read_with_event(self):
         rd, wr = self.make_socketpair()
 
@@ -681,24 +705,81 @@ class ScalableSelectorMixin(object):
         self.assertEqual(limit_nofile // 2, len(s.select()))
 
 
+@skipUnlessHasSelector
+class TestUniqueSelectScenarios(BaseSelectorTestCase):
+    def test_select_module_patched_after_import(self):
+        # This test is to make sure that after import time
+        # calling DefaultSelector() will still give a good
+        # return value. This issue is caused by gevent, eventlet.
+
+        # Now remove all selectors except `select.select`.
+        self.patch_select_module('select')
+
+        # Make sure that the selector returned only uses the selector available.
+        selector = self.make_selector()
+        self.assertIsInstance(selector, selectors.SelectSelector)
+
+    @skipUnlessHasENOSYS
+    def test_select_module_defines_does_not_implement_poll(self):
+        # This test is to make sure that if a platform defines
+        # a selector as being available but does not actually
+        # implement it (kennethreitz/requests#3906) then
+        # DefaultSelector() does not fail.
+
+        # Reset the _DEFAULT_SELECTOR value as if using for the first time.
+        selectors._DEFAULT_SELECTOR = None
+
+        # Now we're going to patch in a bad `poll`.
+        class BadPoll(object):
+            def poll(self, timeout):
+                raise OSError(errno.ENOSYS)
+
+        # Remove all selectors except `select.select` and replace `select.poll`.
+        self.patch_select_module('select', poll=BadPoll)
+
+        selector = self.make_selector()
+        self.assertIsInstance(selector, selectors.SelectSelector)
+
+    @skipUnlessHasENOSYS
+    def test_select_module_defines_does_not_implement_epoll(self):
+        # Same as above test except with `select.epoll`.
+
+        # Reset the _DEFAULT_SELECTOR value as if using for the first time.
+        selectors._DEFAULT_SELECTOR = None
+
+        # Now we're going to patch in a bad `epoll`.
+        def bad_epoll(*args, **kwargs):
+            raise OSError(errno.ENOSYS)
+
+        # Remove all selectors except `select.select` and replace `select.epoll`.
+        self.patch_select_module('select', epoll=bad_epoll)
+
+        selector = self.make_selector()
+        self.assertIsInstance(selector, selectors.SelectSelector)
+
+
 @skipUnless(hasattr(selectors, "SelectSelector"), "Platform doesn't have a SelectSelector")
 class SelectSelectorTestCase(BaseSelectorTestCase):
-    SELECTOR = getattr(selectors, "SelectSelector", None)
+    def setUp(self):
+        self.patch_select_module('select')
 
 
 @skipUnless(hasattr(selectors, "PollSelector"), "Platform doesn't have a PollSelector")
 class PollSelectorTestCase(BaseSelectorTestCase, ScalableSelectorMixin):
-    SELECTOR = getattr(selectors, "PollSelector", None)
+    def setUp(self):
+        self.patch_select_module('poll')
 
 
 @skipUnless(hasattr(selectors, "EpollSelector"), "Platform doesn't have an EpollSelector")
 class EpollSelectorTestCase(BaseSelectorTestCase, ScalableSelectorMixin):
-    SELECTOR = getattr(selectors, "EpollSelector", None)
+    def setUp(self):
+        self.patch_select_module('epoll')
 
 
 @skipUnless(hasattr(selectors, "KqueueSelector"), "Platform doesn't have a KqueueSelector")
 class KqueueSelectorTestCase(BaseSelectorTestCase, ScalableSelectorMixin):
-    SELECTOR = getattr(selectors, "KqueueSelector", None)
+    def setUp(self):
+        self.patch_select_module('kqueue')
 
 
 @skipUnless(hasattr(selectors, "SelectSelector"), "Platform doesn't have a SelectSelector")