+++ CHANGES.rst
@@ -10,6 +10,9 @@ dev (master)
 * Resolved issue where the PyOpenSSL backend would not wrap SysCallError
   exceptions appropriately when sending data. (Pull #1125)
 
+* Selectors now detects a monkey-patched select module after import for modules
+  that patch the select module like eventlet, greenlet. (Pull #1128)
+
 * Reduced memory consumption when streaming zlib-compressed responses
   (as opposed to raw deflate streams). (Pull #1129)